W181: twenty-five more windows cleared, and the VU placeholder macros start coming out of mdlFile.
No new MATCHes this wave. That is the honest headline, and it is the expected
shape of the work: every function touched here was hundreds of instructions
from retail, and the wave's job was to delete work retail never did and to
restore work the placeholders had erased. Over-window count fell 164 -> 139
with nothing lost and nothing newly over.

THE TYPE LENS is now the most productive tool in the project. A 64-bit local
holding a 32-bit call result makes MWCC emit a spurious dsll32/dsra32 pair, and
the cumulative cost of that one mistake is enormous:
  itfPanel 003aa720  1016/960 nd731 -> 960/960 nd18  -- storing FUN_001158b0's
                     result as u32 instead of u64, plus the real 16-element
                     short scratch aggregate instead of fragmented scalars.
  itfPanel 003aaae0  2136/1984 nd1526 -> 1968/1984 nd774, same correction.
  itfPanel 003abb10  1944/1840 -> 1796/1840.
  btlVoice 002eabf0, 002eb2c0, 002ee640, 002efe50, 002f0ea0 -- packet and
                     object returns corrected from u64 to u32, four windows
                     cleared between them.
  y_runtime 00458cb0 -116 bytes, 00457a40 -80, 0045ace0 -92, 0045b190 -28,
                     y_misc 004281f0 cleared -- long/u64 locals to int/u32,
                     one global corrected from u32 to f32.
  mt_evtMain 003646d0 and 00364b10 both cleared when a helper's float return
                     was finally stored in an f32.
  fclCombineMisc 003d74f0 -136, fclCombineEffect 004178c0 cleared,
  fclDrawSprite 0040fb90 -264 -- pointers and counts recovered from u64/long.
A census of every remaining instance is in `build/wip/W182_widelocal.md`:
342 NONMATCHING functions still store a call result in a 64-bit local.

INVENTED ZERO-INITIALIZATION continues to pay. btlFormation 002c5030, 002c5fc0,
002c6300, 002c65d0 and 002c6f50 all shed scratch-array clears the code then
fully populates; three of those windows closed. mt_evtMenu 0037d5d0 cleared on
two invented 12-byte clears immediately before the helper that writes both
buffers.

CONCAT RESIDUE is a third family. `CONCAT11`/`CONCAT22`/`CONCAT31` is raw
Ghidra packing the source should express directly. mt_evtMenu 0037afc0 cleared
its window on one CONCAT22 rewrite; 003709d0 shed 200 bytes; 0037b5d0 shed 24;
frFont 003b2a30 shed 36; and removing all five in fclShopMisc took 003fcad0
from nd176 to nd96 and 003fd030 from nd174 to nd94.

INVENTED DEFENSIVE BRANCHES: btlMain 002a2660 cleared after four impossible
pointer-plus-offset guards came out; mt_evtMenu 00372c40 lost a null-state
diagnostic guard retail does not have.

THE mdlFile VU UNLOCK BEGINS. mdlFile.c defines ~30 placeholder macros
(`_lqc2`, `_sqc2`, `_qmtc2`, `_vmul`, `_vadd`, `_vmadd`, `_vitof`, `_vftoi0`,
`_vrsqrt`, ...) that all expand to `mdlVecZero()` or `mdlVecKeep()`. They emit
scalar garbage, so the 41 NONMATCHING functions calling them -- 1961 call sites
-- were unmatchable by construction. Three agents converted the smaller ones to
real fixed-register EE inline asm, decoding each instruction from the retail
raw opcode rather than guessing a mnemonic. This is legitimate hardware asm:
VU macro-mode is genuinely inexpressible in C, effMisc.c is the in-repo
precedent, and nine in-scope functions already MATCH with VU in their window.
  00329ed0  284/592 -> 580/592   (the pathfinder; `d86a0000` decoded as
                                  lqc2 vf10,0(base), sqrtf for retail's sqrt.s)
  00325c10  192/336 -> 312/336
  00325e40  332/496 -> 464/496
  00337fd0  552/912 -> 872/912
  00338ea0  532/864 -> 860/864
  00326280  380/304 -> 308/304   (76 bytes over down to 4)
  00335180 and 003377f0, an exact 624/624 nd479 twin pair, both to 556 nd389
  00336630  window cleared
The six densest functions -- 00333970 (100 sites), 00336e80 (190), 003399a0,
0033a5b0, 003407e0, 00335a70 -- were correctly left alone: they carry fixed VU
register state across calls and need whole-function reconstruction, not a
partial conversion that would leave the source half-truthful.

The working recipe is recorded in the W181VuLow transcript: decode the retail
raw words first; bracket each block with `.set noreorder` / `.set reorder`; use
typed 16-byte byte arrays for sqc2 destinations with a `memory` clobber and
exact VU/GPR/FPR clobbers; split blocks at C branches and call boundaries so VU
register state stays live; use honest 16-byte aggregate copies where retail
uses LQ/SQ; use `sqrtf` where retail has `sqrt.s`.

Checkpoint: 6421 / 8211 first-party MATCH (78.2%). Zero lost, zero new
over-window, zero compile errors, zero stale markers.

+++ src/Main/Social/sfl_result.c
@@ -2397,7 +2397,7 @@ int func_001fc870(u8* event, void* target)
 
     raw = FUN_0030bc50(target);
     if (raw < 0)
-        amount = (f32)(((u32)raw >> 1) | (raw & 1)) * 2.0f;
+        amount = (f32)(s32)(((u32)raw >> 1) | (raw & 1)) * 2.0f;
     else
         amount = (f32)raw;
